{{cookiecutter.github_repository_name}}/requirements.txt
# Core
pytz==2019.1  # django dependency
Django==2.2.2

# Storages
psycopg2-binary==2.7.7

# Rest apis
djangorestframework==3.9.4
django-filter==2.1.0
six==1.12.0  # django-rest-auth pytest dependency
django-rest-auth==0.9.5

# AWS
docutils==0.14  # botocore dependency
jmespath==0.9.4  # boto3 botocore dependency
boto3==1.9.166
botocore==1.12.166
s3transfer==0.2.1

# Dynamic Documentations
certifi==2019.3.9  # coreapi dependency
coreschema==0.0.4  # coreapi dependency
itypes==1.1.0  # coreapi dependency
uritemplate==3.0.0  # coreapi dependency
coreapi==2.3.3
